requirements.txt
# Xcleaners - Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.7

# Database
asyncpg==0.29.0

# Cache / Rate Limiting
redis[hiredis]==5.0.0

# Security
bcrypt==4.1.2
PyJWT==2.8.0
cryptography>=43.0.0

# AI Providers
anthropic==0.40.0
openai==1.0.0

# Voice
edge-tts==6.1.0

# HTTP Client
httpx==0.26.0

# Email
resend==2.0.0

# Image Processing
Pillow>=11.0.0

# Push Notifications
pywebpush==1.14.1

# Payments
stripe==7.0.0

# SMS (Twilio for Xcleaners notifications)
twilio==8.0.0

# Environment
python-dotenv==1.0.0

# Validation
email-validator==2.1.0

# Scheduler
apscheduler==3.10.0

# SSE (Server-Sent Events)
sse-starlette==1.6.0

# Google Analytics & Search Console
google-analytics-data==0.18.0
google-api-python-client==2.100.0
google-auth==2.23.0
